requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
validate-docbr==1.9.0
cryptography==42.0.8